Renamed chainer filename
Prefer hyphen names for file names
