+++ src/endpoint-generator.js
